solved the sign extension puzzle

+++ include/libpltnc/byteswap.h
@@ -8,7 +8,7 @@
 #include "libpltnc/types.h"
 
 #ifndef __ASSEMBLER__
-	u32 swap32(u32 val);
+	u64 swap32(u64 val);
 	u64 swap64(u64 val);
 	u64 has_zero_byte(u64 val);
 #endif
